+++ example/test/add.spec.js
@@ -1,11 +1,9 @@
-/* eslint-env mocha */
-
-const assert = require('assert')
-const add = require('../src/add')
+import assert from 'node:assert';
+import add from '../src/add.js';
 
 describe('add()', () => {
   it('should return sum', () => {
-    const sum = add(1, 2)
-    assert.strictEqual(sum, 3)
-  })
-})
+    const sum = add(1, 2);
+    assert.strictEqual(sum, 3);
+  });
+});