+++ src/index.js
@@ -1,9 +1,8 @@
-/*
-** based on https://github.com/mochajs/mocha/blob/master/lib/reporters/json.js
-*/
+// based on https://github.com/mochajs/mocha/blob/master/lib/reporters/json.js
+/* eslint-disable n/no-sync -- Sync API */
 
-const fs = require('fs')
-const Mocha = require('mocha')
+import fs from 'node:fs';
+import mocha from 'mocha';
 
 const {
   EVENT_TEST_PASS,
@@ -11,63 +10,88 @@ const {
   EVENT_TEST_END,
   EVENT_RUN_END,
   EVENT_TEST_PENDING
-} = Mocha.Runner.constants
+} = mocha.Runner.constants;
 
-const DEFAULT_REPORT_PATH = 'report.json'
+const DEFAULT_REPORT_PATH = 'report.json';
 
-function Reporter (runner, options) {
-  Mocha.reporters.Base.call(this, runner, options)
-  const self = this
+/* eslint-disable jsdoc/imports-as-dependencies -- Ok */
+/**
+ *
+ */
+class Reporter extends mocha.reporters.Base {
+  /**
+   * @param {import('mocha').Runner} runner
+   * @param {{
+   *   reporterOptions?: {output: string}
+   * }} options
+   */
+  constructor (runner, options) {
+    /* eslint-enable jsdoc/imports-as-dependencies -- Ok */
 
-  const tests = []
-  const pending = []
-  const failures = []
-  const passes = []
+    super(runner, options);
 
-  runner.on(EVENT_TEST_END, function (test) {
-    tests.push(test)
-  })
+    const tests = [];
+    const pending = [];
+    const failures = [];
+    const passes = [];
 
-  runner.on(EVENT_TEST_PASS, function (test) {
-    passes.push(test)
-  })
+    runner.on(EVENT_TEST_END, (test) => {
+      tests.push(test);
+    });
 
-  runner.on(EVENT_TEST_FAIL, function (test) {
-    failures.push(test)
-  })
+    runner.on(EVENT_TEST_PASS, (test) => {
+      passes.push(test);
+    });
 
-  runner.on(EVENT_TEST_PENDING, function (test) {
-    pending.push(test)
-  })
+    runner.on(EVENT_TEST_FAIL, (test) => {
+      failures.push(test);
+    });
 
-  runner.once(EVENT_RUN_END, function () {
-    const obj = {
-      stats: self.stats,
-      tests: tests.map(clean),
-      pending: pending.map(clean),
-      failures: failures.map(clean),
-      passes: passes.map(clean)
-    }
-    runner.testResults = obj
-    const json = JSON.stringify(obj, null, 2)
-    let path = DEFAULT_REPORT_PATH
-    const { reporterOptions } = options
-    if (reporterOptions) {
-      const { output } = reporterOptions
-      if (output) {
-        path = output
+    runner.on(EVENT_TEST_PENDING, (test) => {
+      pending.push(test);
+    });
+
+    runner.once(EVENT_RUN_END, () => {
+      const obj = {
+        stats: this.stats,
+        tests: tests.map((test) => clean(test)),
+        pending: pending.map((element) => clean(element)),
+        failures: failures.map((failure) => clean(failure)),
+        passes: passes.map((pass) => clean(pass))
+      };
+      runner.testResults = obj;
+      const json = JSON.stringify(obj, null, 2);
+      let path = DEFAULT_REPORT_PATH;
+      const {reporterOptions} = options;
+      if (reporterOptions) {
+        const {output} = reporterOptions;
+        if (output) {
+          path = output;
+        }
       }
-    }
-    const out = fs.openSync(path, 'w')
-    fs.writeSync(out, json)
-    fs.closeSync(out)
-  })
+      const out = fs.openSync(path, 'w');
+      fs.writeSync(out, json);
+      fs.closeSync(out);
+    });
+  }
 }
 
+/* eslint-disable jsdoc/imports-as-dependencies -- Ok */
+/**
+ * @param {import('mocha').Test} test
+ * @returns {{
+ *   title: string,
+ *   fullTitle: string,
+ *   duration: number,
+ *   currentRetry: boolean,
+ *   err: Error | object
+ * }}
+ */
 function clean (test) {
-  let err = test.err || {}
-  if (err instanceof Error) {
-    err = errorJSON(err)
+  /* eslint-enable jsdoc/imports-as-dependencies -- Ok */
+  let err = test.err || {};
+  if (err && typeof err === 'object' && typeof err.message === 'string') {
+    err = errorJSON(err);
   }
 
   return {
@@ -76,30 +100,38 @@ function clean (test) {
     duration: test.duration,
     currentRetry: test.currentRetry(),
     err: cleanCycles(err)
-  }
+  };
 }
 
+/**
+ * @param {object} obj
+ * @returns {object}
+ */
 function cleanCycles (obj) {
-  const cache = []
+  const cache = [];
   return JSON.parse(
     JSON.stringify(obj, function (key, value) {
       if (typeof value === 'object' && value !== null) {
-        if (cache.indexOf(value) !== -1) {
-          return '' + value
+        if (cache.includes(value)) {
+          return String(value);
         }
-        cache.push(value)
+        cache.push(value);
       }
-      return value
+      return value;
     })
-  )
+  );
 }
 
+/**
+ * @param {Error} err
+ * @returns {Record<string, string>}
+ */
 function errorJSON (err) {
-  const res = {}
-  Object.getOwnPropertyNames(err).forEach(function (key) {
-    res[key] = err[key]
-  }, err)
-  return res
+  const res = {};
+  Object.getOwnPropertyNames(err).forEach((key) => {
+    res[key] = err[key];
+  });
+  return res;
 }
 
-module.exports = Reporter
+export default Reporter;